+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.1.0
 
-- Adding total time, projected time, and estimated savings/loss information. Requires moon >= 0.15.
+- Added total time, projected time, and estimated savings/loss information. Requires moon >= 0.15.
+- Added `slow-threshold` input setting.
 
 # 1.0.0